overload_handler.py: reduce api calls
Considering the API call limitations, we adopt a more
conservative number of calls by allowing only 5 consecutive
API calls.
If charging amps is set to max we also end the supervised session.
Because many times the peaks are transitory and take less than 2min
we wait for 10*sleep_time with MAX_AMPS set to half of the value
without making any changes to charging_amps, hopefully the peak will
pass and with only one more call we can get back to max and end
the supervised session.